delta-webapp/requirements.txt
asgiref==3.5.2
biopython==1.80
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.3
ete3==3.1.2
gunicorn==20.1.0
idna==3.4
itolapi==4.1.1
Jinja2==3.1.2
llvmlite==0.39.1
MarkupSafe==2.1.1
numba==0.56.4
numpy==1.23.5
pandas==1.5.1
pastml==1.9.40
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
scipy==1.9.3
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.14
whitenoise==6.2.0
